+++ config/services/core/main.yaml
@@ -3,6 +3,7 @@ imports:
   - { resource: 'third_party.yaml' }
   - { resource: 'templates.yaml' }
   - { resource: 'new_subscription_stats.yaml' }
+  - { resource: 'subscription_movement_stats.yaml' }
 
 services:
   # default configuration for services in *this* file
