+++ config/services/core/subscription_movement_stats.yaml
@@ -1,0 +1,13 @@
+services:
+  # Interface binding
+  BillaBear\Repository\Stats\SubscriptionMovementStatsRepositoryInterface: '@BillaBear\Repository\SubscriptionMovementStatsRepository'
+
+  # Service repository
+  BillaBear\Repository\SubscriptionMovementStatsRepository:
+    arguments:
+      $entityRepository: '@app.repository.orm.subscription_movement_stats'
+      $entityManager: '@doctrine.orm.entity_manager'
+
+  # ORM repository
+  app.repository.orm.subscription_movement_stats:
+    class: BillaBear\Repository\Orm\SubscriptionMovementStatsRepository
